Remove GUI system tests.
When writing tests for the new controls in #30, it was found that a failing test would cause the framework to crash when starting the next test, and not produce any output whatsoever.
It seems that when one test fails an assertion, it failed to correctly shut down the window,
leading to the next test being unable to open it's window, and the program never progressing.
I suspect this may be due to the swapping of QQuickView for QQmlApplicationEngine in 42c232f8fa9becd21493af0b0434b64195da3738,
as the QQAE documentation (https://doc.qt.io/archives/qt-5.7/qqmlapplicationengine.html)
mentions that it changes the behaviour of quit.
Swapping back to QQuickView is not practical, as it doesn't support window menu bars.
Due to this, and the general difficulty working with PySide2/QML/pytest-qt, we've decided to drop these tests, and to add a manual test document covering the steps that should be done to verify the UI before any releases.